requirements.txt
fastapi
uvicorn
sse-starlette
python-dotenv
langchain
langchain-openai
openai
supabase
